Рабочая версия сервер(1 из 10 раз падают рандомные тесты???), не проходит 1 тест в статс

+++ ewm/src/main/java/ru/practicum/stats/EwmViewStats.java
@@ -12,5 +12,5 @@ public class EwmViewStats {
 
     String uri;
 
-    Integer hits;
+    Long hits;
 }
